Ask for input values and handle exceptions
When it comes to interaction with the DB it will now ask the user for
credentials and show the default values. In case the user enters for the
temporalily used table a name that already exists, the exception will be
captured and the user may enter a valid tablename.

These changes are made, because it is way more comfortable for a user to
enter its credentials via I/O instead of changing the source files.
If the database raises a trivial exception, like that the table name
already exists, the program should not end abruptly.

For Exception handling Control.Exception has been used.