Use streams for XML parsing
We use `conduit` in order to stream the input of the XML file, parse it
and upsert the entry in the database. This operation takes about 30
minutes on i5 intel Core with 8GB RAM for a table with 10 Mio entries
and about 1.5 Mio updates for rows.

The first approach to parse the XML file has used the Text.XML.Cursor library
has been rejected since the whole data of the XML needs to stay in
memory. The current approach does not face this problem.