Use copy command to create tmp table and upsert it
The parsed XML file is saved as csv file. Afterwards, a temporary table
is created which is filled with the entries in the csv file by the copy
command. With an Update/Insert query the temporary table is merges with
'person'.
As a consequence, the module for parsing the xml file is a bit less
fancy. It reads in the file, creates a person (which is not that
necessary, but more functional) and writes the input back to a csv file.

This approach works a lot of faster. The first approach took a similar
route, but created a (non-temporary) table by using many 'insert'
queries. 'Copy' on a temporary table is a lot faster, which totally
makes sense.